# Parse C# attributes to determine element attributes