--- a/Directory.Build.targets
+++ b/Directory.Build.targets
@@ -10,6 +10,7 @@
     FEATURE_CLOSE                    : Stream.Close is available in the platform
     FEATURE_DCS_SETTINGS             : DataContractSerializer settings class exists
     FEATURE_PACKAGE_FLUSH            : System.IO.Package.Flush will correctly flush all data
+    FEATURE_NO_CONDITIONAL_WEAK_TABLE: ConditionalWeakTable is not available
     FEATURE_SERIALIZATION            : Serialization using ISerializable is in the platform
     FEATURE_SCHEMA_GENERATOR         : Provides access to helpers to generate schema information
     FEATURE_SYSTEMEXCEPTION          : SytemException is available in the platform
@@ -20,7 +21,7 @@
   -->
 
   <PropertyGroup Condition=" '$(TargetFramework)' == 'net35' ">
-    <DefineConstants>$(DefineConstants);FEATURE_PACKAGE_FLUSH;FEATURE_XML_PROHIBIT_DTD</DefineConstants>
+    <DefineConstants>$(DefineConstants);FEATURE_PACKAGE_FLUSH;FEATURE_XML_PROHIBIT_DTD;FEATURE_NO_CONDITIONAL_WEAK_TABLE</DefineConstants>
   </PropertyGroup>
 
   <PropertyGroup Condition=" '$(TargetFramework)' == 'net35' OR '$(TargetFramework)' == 'net40' OR '$(TargetFramework)' == 'net452' ">