--- a/Open-XML-SDK.sln
+++ b/Open-XML-SDK.sln
@@ -8,6 +8,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		appveyor.yml = appveyor.yml
 		Directory.Build.props = Directory.Build.props
 		Directory.Build.targets = Directory.Build.targets
+		DocumentFormat.OpenXml.Package.props = DocumentFormat.OpenXml.Package.props
 		GitVersion.yml = GitVersion.yml
 		LICENSE = LICENSE
 		NuGet.Config = NuGet.Config