--- a/src/DocumentFormat.OpenXml/AttributeTagCollection.cs
+++ b/src/DocumentFormat.OpenXml/AttributeTagCollection.cs
@@ -4,29 +4,18 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
 
 namespace DocumentFormat.OpenXml
 {
-    [DebuggerDisplay("Length = {Length}")]
-    [DebuggerTypeProxy(typeof(AttributeTagCollectionDebugView))]
     internal readonly struct AttributeTagCollection : IEnumerable<OpenXmlAttribute>
     {
         private readonly ReadOnlyArray<AttributeTag> _tags;
-        private readonly OpenXmlSimpleType[] _fields;
+        private readonly OpenXmlElement _element;
 
-        public AttributeTagCollection(in ReadOnlyArray<AttributeTag> tags)
+        public AttributeTagCollection(OpenXmlElement element, in ReadOnlyArray<AttributeTag> tags)
         {
             _tags = tags;
-
-            if (tags.Length == 0)
-            {
-                _fields = Cached.Array<OpenXmlSimpleType>();
-            }
-            else
-            {
-                _fields = new OpenXmlSimpleType[tags.Length];
-            }
+            _element = element;
         }
 
         public bool Any() => Length > 0;
@@ -35,15 +24,15 @@ namespace DocumentFormat.OpenXml
 
         public AttributeEntry this[string namespaceUri, string tagName] => this[GetIndex(namespaceUri, tagName)];
 
-        public int Length => _fields.Length;
+        public int Length => _tags.Length;
 
         public Enumerator GetEnumerator() => new Enumerator(this);
 
         private int GetIndex(string namespaceUri, string tagName)
         {
-            if (!string.IsNullOrEmpty(tagName) && namespaceUri != null && NamespaceIdMap.TryGetNamespaceId(namespaceUri, out byte nsId))
+            if (!string.IsNullOrEmpty(tagName) && namespaceUri != null && NamespaceIdMap.TryGetNamespaceId(namespaceUri, out var nsId))
             {
-                for (int i = 0; i < _tags.Length; i++)
+                for (var i = 0; i < _tags.Length; i++)
                 {
                     var tag = _tags[i];
 
@@ -97,13 +86,10 @@ namespace DocumentFormat.OpenXml
             public bool MoveNext() => ++_index < _collection.Length;
         }
 
-        [DebuggerDisplay("{Tag}: {Value}")]
         public readonly struct AttributeEntry
         {
-            [DebuggerBrowsable(DebuggerBrowsableState.Never)]
             private readonly AttributeTagCollection _collection;
 
-            [DebuggerBrowsable(DebuggerBrowsableState.Never)]
             private readonly int _index;
 
             public AttributeEntry(in AttributeTagCollection collection, int index)
@@ -116,22 +102,11 @@ namespace DocumentFormat.OpenXml
 
             public ref readonly AttributeTag Tag => ref _collection._tags[_index];
 
-            public ref OpenXmlSimpleType Value => ref _collection._fields[_index];
+            public OpenXmlSimpleType Value => _collection._tags[_index].GetValue(_collection._element);
+
+            public void SetValue(OpenXmlSimpleType value) => _collection._tags[_index].SetValue(_collection._element, value);
 
             public bool HasValue => Value != null;
-        }
-
-        internal sealed class AttributeTagCollectionDebugView
-        {
-            private readonly AttributeTagCollection _collection;
-
-            public AttributeTagCollectionDebugView(AttributeTagCollection collection)
-            {
-                _collection = collection;
-            }
-
-            [DebuggerBrowsable(DebuggerBrowsableState.RootHidden)]
-            public AttributeEntry[] Items => _collection.ToArray();
         }
     }
 }