--- a/src/DocumentFormat.OpenXml/AttributeTagCollection.cs
+++ b/src/DocumentFormat.OpenXml/AttributeTagCollection.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 
 #if FEATURE_NO_CONDITIONAL_WEAK_TABLE
 using TagLookup = DocumentFormat.OpenXml.LockingDictionary<System.Type, DocumentFormat.OpenXml.AttributeTag[]>;