--- a/src/DocumentFormat.OpenXml/Cached.cs
+++ b/src/DocumentFormat.OpenXml/Cached.cs
@@ -26,26 +26,5 @@ namespace DocumentFormat.OpenXml
         {
             internal static ReadOnlyCollection<T> Value = new ReadOnlyCollection<T>(Array<T>());
         }
-
-        public static Func<T> Activator<T>() where T : new() => CachedActivator<T>.Activator;
-
-        private static class CachedActivator<T>
-            where T : new()
-        {
-            private static readonly Expression<Func<T>> Creator = () => new T();
-
-            public static Func<T> Activator { get; } = Creator.Compile();
-        }
-
-        public static Func<TBase> Activator<TInstance, TBase>()
-            where TInstance : TBase, new() => CachedActivator<TInstance, TBase>.Activator;
-
-        private static class CachedActivator<TInstance, TBase>
-            where TInstance : TBase, new()
-        {
-            private static readonly Expression<Func<TBase>> Creator = () => new TInstance();
-
-            public static Func<TBase> Activator { get; } = Creator.Compile();
-        }
     }
 }