--- a/src/DocumentFormat.OpenXml/DocumentFormat.OpenXml.csproj
+++ b/src/DocumentFormat.OpenXml/DocumentFormat.OpenXml.csproj
@@ -21,9 +21,9 @@
     <!-- Remove all files under System as they need to be specifically opted into. They are
          helper files to fill the gaps of the various platforms that are targeted -->
     <Compile Remove="System\**\*.cs" />
+    <Compile Include="System\Collections\Generic\LockingDictionary.cs" />
     <Compile Include="System\UriHelper.cs" />
     <Compile Remove="PartActivator.Custom.cs" />
-    <Compile Condition="$(DefineConstants.Contains('FEATURE_NO_CONDITIONAL_WEAK_TABLE'))" Include="System\Collections\Generic\LockingDictionary.cs" />
   </ItemGroup>
 
   <Choose>