This change is to reintroduce `SchemaAttrAttribute` usage but rely on it instead of manually building the table of attributes. This allows much more concise schema definitions and actually appears to be faster in the benchmarks we have. This is part of the change to make the schema more extensible and modernize the validation system.

** Performance diff **

This branch:

```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i7-6600U CPU 2.60GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=2.1.503
  [Host]     : .NET Core 1.1.10 (CoreCLR 4.6.26906.01, CoreFX 4.6.24705.01), 64bit RyuJIT
  Job-BZAUSC : .NET Core 1.1.10 (CoreCLR 4.6.26906.01, CoreFX 4.6.24705.01), 64bit RyuJIT

Runtime=Core  
```

|        Method|         Mean|       Error|      StdDev|     Gen 0|    Gen 1| Gen 2| Allocated |
|-|-|-|-|-|-|-|-|
|        Create|    267.09 us|  27.8643 us|  82.1584 us|      36.6211|        -| |  75.45 KB |
|  CreateNoSave|     31.45 us|   0.4830 us|   0.4033 us|       9.3994|        -| |  19.38 KB |
|      ReadFile|  1,324.69 us|  11.6884 us|  10.9334 us|    269.5313|  52.7344| | 615.76 KB |
|  Validation|  52.13 ms|  1.312 ms|  1.095 ms|  3937.5000|  1250.0000|  500.0000|   15.94 MB |


MASTER:

```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i7-6600U CPU 2.60GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=2.1.503
  [Host]     : .NET Core 1.1.10 (CoreCLR 4.6.26906.01, CoreFX 4.6.24705.01), 64bit RyuJIT
  Job-QBMVLR : .NET Core 1.1.10 (CoreCLR 4.6.26906.01, CoreFX 4.6.24705.01), 64bit RyuJIT

Runtime=Core  
```

|        Method|         Mean|       Error|      StdDev|     Gen 0|    Gen 1| Gen 2| Allocated |
|-|-|-|-|-|-|-|-|
|        Create|           NA|          NA|          NA|       N/A|      N/A| |       N/A |
|  CreateNoSave|     69.25 us|    3.339 us|    9.844 us|    9.3994|        -| |   19840 B |
|      ReadFile|  3,209.37 us|  186.967 us|  551.276 us|  269.5313|  50.7813| |  630536 B |
|  Validation|  74.62 ms|  3.697 ms|  10.90 ms|   3937.5000|  1250.0000|  500.0000|   15.94 MB |